chore: moved some types around
