+++ libs/icon-button/src/types.ts
@@ -1,0 +1,10 @@
+import { IconButtonProps as ChakraIconButtonProps, TooltipProps } from '@chakra-ui/react';
+
+export interface IconButtonProps extends Omit<ChakraIconButtonProps, 'aria-label'> {
+  'aria-label'?: string;
+  reversed?: boolean;
+  testId: string;
+  tooltipProps?: Omit<TooltipProps, 'children'> & {
+    label: string;
+  };
+}
